Refactor of the project to support ANT build system
This allows a zip file to consistently be produced using ANT. To trigger simply run the default target
The build will output a software-license-manager.zip file in the subdirectory `target/dist`.  The build removes composer relater configuration files as well as *.po files which are not required for language domain.  It creates a readme.txt and the main php file using text replacement using build.properties.

The creates a slimmed down zip without the cruft associated with the development environment.  Additionally a number of index.html files are added for security.  Vscode and other ide specific files are excluded via .gitignore.